fix(web): keep dashboard widget components out of the appearance store's import graph
The portal imports the app's appearance store. Since #1191 that store
imported widget-registry.ts for DEFAULT_TOP_WIDGETS, and the registry's
dynamic loaders pull every dashboard widget, and their $api imports, into
the portal bundle. The portal has no alias for those paths, so its build
has failed on main since 2026-09-06 with "Rollup failed to resolve import
$api/generated/batteries.generated.remote".

The ids now live in top-widget-ids.ts with no component behind them:
TOP_WIDGET_IDS, TopWidgetId, isTopWidgetId, LOADABLE_TOP_WIDGETS,
DEFAULT_TOP_WIDGETS and knownTopWidgets. widget-registry.ts re-exports
them and types its loader map as Record<TopWidgetId, WidgetLoader>, so an
id without a loader, or a loader without an id, still fails to compile.
The store imports the ids module through a relative path, which the
portal can resolve.

Behaviour is unchanged: the id set, its order and the defaults are the
same nine and three; isTopWidgetId still rejects inherited object keys
("toString", "__proto__") because it consults a Set rather than the
loader object.
